Delete useless code and add utility.h file.
