+++ api/envoy/config/filter/network/dubbo_proxy/v2alpha1/route.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 
 package envoy.config.filter.network.dubbo_proxy.v2alpha1;
-option java_package = "io.envoyproxy.envoy.config.filter.network.thrift_proxy.v2alpha1";
+option java_package = "io.envoyproxy.envoy.config.filter.network.dubbo_proxy.v2alpha1";
 option java_multiple_files = true;
 option go_package = "v2";
 
@@ -48,9 +48,6 @@ message MethodMatch {
   message Parameter {
     // Parameter index, starting from 0.
     uint32 index = 1;
-
-    // Parameter types, only basic data types are supported.
-    string type = 2;
 
     oneof parameter_match_specifier {
       // If specified, header match will be performed based on the value of the header.
